+++ src/json_j.mli
@@ -98,8 +98,11 @@ type ghost_abilities = Json_t.ghost_abilities = {
   mutable mothwing_cloak: bool;
   mutable shade_cloak: bool;
   mutable vengeful_spirit: bool;
+  mutable shade_soul: bool;
   mutable desolate_dive: bool;
-  mutable howling_wraiths: bool
+  mutable descending_dark: bool;
+  mutable howling_wraiths: bool;
+  mutable abyss_shriek: bool
 }
 
 type game_progress = Json_t.game_progress = {