+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - add respawn targets for hazards
 - fix double-pogoing
 - add multi-hits for Vengeful Spirit and Howling Wraiths
+- add spell upgrades
 
 ## [0.0.8] - 2023-11-19